[REFACTOR] #251: context 이름 수정
