+++ frontend/src/containers/TrimPage/TrimBannerContainer.tsx
@@ -6,7 +6,7 @@ import { IMG_URL } from '../../utils/apis';
 import CenterWrapper from '../../components/layout/CenterWrapper';
 import Banner from '../../components/common/banner/Banner';
 import HmgTag from '../../components/common/hmgTag/HmgTag';
-import { ICartype, TrimContext } from '../../context/TrimContext';
+import { ICartype, TrimContext } from '../../context/TrimProvider';
 import Loading from '../../components/loading/Loading';
 
 export default function TrimBannerContainer() {
