+++ frontend/src/context/TrimProvider.tsx
